requirements.txt
biopython>=1.83
python-dotenv>=1.0.0
httpx>=0.25.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0